move assignments logic to separate class
